+++ src/common/CurrencyEngine.ts
@@ -808,6 +808,13 @@ export class CurrencyEngine<
     return checkpointB
   }
 
+  updateTransactionEvents(): void {
+    if (this.transactionEvents.length > 0) {
+      this.currencyEngineCallbacks.onTransactions(this.transactionEvents)
+      this.transactionEvents = []
+    }
+  }
+
   // Called by EthereumNetwork
   updateOnAddressesChecked(): void {
     if (this.addressesChecked) {