+++ src/cosmos/engine/CosmosEngine.ts
@@ -553,11 +553,7 @@ export class CosmosEngine extends CurrencyEngine<
     }
     this.otherData.archivedTxLastCheckTime = Date.now()
     this.walletLocalDataDirty = true
-
-    if (this.transactionEvents.length > 0) {
-      this.currencyEngineCallbacks.onTransactions(this.transactionEvents)
-      this.transactionEvents = []
-    }
+    this.updateTransactionEvents()
   }
 
   async queryTransactionsInner(